BACK - Story/104697 - RELEASE

+++ SME.CDEP.Dominio/Entidades/AcervoTridimensionalCompleto.cs
@@ -13,10 +13,10 @@ public class AcervoTridimensionalCompleto: EntidadeBaseAuditavel
     public long ConservacaoId { get; set; }
     public long Quantidade { get; set; }
     public string Descricao { get; set; }
-    public float? Largura { get; set; }
-    public float? Altura { get; set; }
-    public float? Profundidade { get; set; }
-    public float? Diametro { get; set; }
+    public double? Largura { get; set; }
+    public double? Altura { get; set; }
+    public double? Profundidade { get; set; }
+    public double? Diametro { get; set; }
     public ArquivoResumido[] Arquivos  { get; set; }
     public long ArquivoId { get; set; }
     public string ArquivoNome { get; set; }
